Move class Album from botogram.objects.media to botogram.objects.mixins
