+++ botogram/objects/__init__.py
@@ -22,11 +22,11 @@
 
 from .chats     import User, Chat, UserProfilePhotos, Permissions
 from .media     import PhotoSize, Photo, Audio, Voice, Document, Sticker, \
-                       Video, Contact, Location, Venue, Album
+                       Video, Contact, Location, Venue
 from .messages  import Message
 from .markup    import ReplyKeyboardMarkup, ReplyKeyboardHide, ForceReply
 from .updates   import Update, Updates
-
+from .mixins import Album
 
 __all__ = [
     # Chats-related objects
